refactor: 모달 관련 로직을 modalView로 이동 및 movieListView에서 제거
